backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
jinja2==3.1.5
python-multipart==0.0.20
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
httpx==0.28.1
python-dotenv==1.0.1
cryptography==44.0.0
